requirements.txt
certifi==2024.2.2
charset-normalizer==2.0.12
idna==3.6
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2024.1
PyYAML==6.0.1
requests==2.27.1
six==1.16.0
urllib3==1.26.18
